# docs/swagger.yml
basePath: /api/v1
consumes:
    - application/json
definitions:
    BaseResponseBody:
        description: Contains error and message fields.
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
        required:
            - error
            - message
        title: BaseResponseBody represents the base response body for API.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
...
    GetUserResponseBody:
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            user:
                $ref: '#/definitions/User'
        required:
            - error
            - message
            - user
        title: GetUserResponseBody includes the signle user for a given ID.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
    GetUsersResponseBody:
        properties:
            count:
                format: int64
                type: integer
                x-go-name: Count
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            users:
                items:
                    $ref: '#/definitions/User'
                type: array
                x-go-name: Users
        required:
            - error
            - message
            - count
            - users
        title: GetUsersResponseBody includes the slice of users.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
    LoginRequestBody:
        properties:
            email:
                maxLength: 255
                minLength: 6
                type: string
                x-go-name: Email
            password:
                maxLength: 256
                minLength: 8
                type: string
                x-go-name: Password
        required:
            - email
            - password
        title: LoginRequestBody includes the email and password of the user.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/parameters
...
    RegisterLoginUserResponseBody:
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            token:
                type: string
                x-go-name: Token
            user:
                $ref: '#/definitions/User'
        required:
            - error
            - message
            - token
            - user
        title: RegisterLoginUserResponseBody includes the token and user model when user register or login.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
    RegisterRequestBody:
        properties:
            email:
                maxLength: 255
                minLength: 6
                type: string
                x-go-name: Email
            password:
                maxLength: 256
                minLength: 8
                type: string
                x-go-name: Password
            username:
                maxLength: 32
                minLength: 1
                type: string
                x-go-name: Username
        required:
            - email
            - username
            - password
        title: RegisterRequestBody includes the email, username, and password of the user being created.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/parameters
...
    User:
        description: User represents the user for this application
        properties:
            avatar_url:
                description: URL to user avatar
                type: string
                x-go-name: AvatarURL
            created_at:
                description: The time the user was registered
                format: date-time
                type: string
                x-go-name: CreatedAt
            email:
                description: The email for this user
                type: string
                x-go-name: Email
            id:
                description: The id for this user
                format: uuid
                type: string
                x-go-name: ID
            name:
                description: The name for this user
                type: string
                x-go-name: Name
            updated_at:
                description: The time the user was updated
                format: date-time
                type: string
                x-go-name: UpdatedAt
            username:
                description: The username for this user
                type: string
                x-go-name: Username
        required:
            - id
            - email
            - username
            - created_at
            - updated_at
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/models
    UserUpdateRequestBody:
        properties:
            email:
                maxLength: 255
                minLength: 6
                type: string
                x-go-name: Email
            name:
                maxLength: 32
                type: string
                x-go-name: Name
            password:
                maxLength: 256
                minLength: 8
                type: string
                x-go-name: Password
            username:
                maxLength: 32
                minLength: 1
                type: string
                x-go-name: Username
        title: UserUpdateRequestBody includes the new email, username, name or password for the user.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/parameters
host: localhost:3040
info:
    title: Diverse API.
    version: 0.0.1
paths:
...
    /fetch:
        get:
            description: Get user and new token if user exists
            operationId: fetchUser
            responses:
                "200":
                    $ref: '#/responses/RegisterLoginUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - User
    /login:
        post:
            description: Returns the user and token by given credentials
            operationId: loginUser
            parameters:
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/LoginRequestBody'
            responses:
                "200":
                    $ref: '#/responses/RegisterLoginUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            tags:
                - User
...
    /register:
        post:
            description: Returns the user and token by given credentials
            operationId: createUser
            parameters:
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/RegisterRequestBody'
            responses:
                "201":
                    $ref: '#/responses/RegisterLoginUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            tags:
                - User
    /users:
        get:
            description: Returns a list of all users
            operationId: allowempty
            responses:
                "200":
                    $ref: '#/responses/GetUsersResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            tags:
                - User
    /users/{user}:
        delete:
            description: Delete user by id
            operationId: deleteUser
            parameters:
                - format: uuid
                  in: path
                  name: user
                  required: true
                  type: string
                  x-go-name: User
            produces:
                - application/json
            responses:
                "204":
                    $ref: '#/responses/DeleteUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            schemes:
                - http
                - https
            security:
                - bearerAuth: []
            tags:
                - User
        get:
            description: Returns the user by given id
            operationId: getUser
            parameters:
                - format: uuid
                  in: path
                  name: user
                  required: true
                  type: string
                  x-go-name: User
            responses:
                "200":
                    $ref: '#/responses/GetUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            tags:
                - User
        patch:
            description: Update user by id with given fields
            operationId: updateUser
            parameters:
                - format: uuid
                  in: path
                  name: user
                  required: true
                  type: string
                  x-go-name: User
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/UserUpdateRequestBody'
            responses:
                "201":
                    $ref: '#/responses/UpdateUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - User
...
produces:
    - application/json
responses:
...
    DeleteUserResponse:
        description: DeleteUserResponse represents response for successfully delete user request.
    ErrorResponse:
        description: |-
            ErrorResponse represents the error response body for API.
            Contains error and message fields.
        schema:
            $ref: '#/definitions/BaseResponseBody'
...
    GetUserResponse:
        description: GetUserResponse represent the response retrived on get user request.
        schema:
            $ref: '#/definitions/GetUserResponseBody'
    GetUsersResponse:
        description: GetUsersResponse represent the response retrived on get users request.
        schema:
            $ref: '#/definitions/GetUsersResponseBody'
    RegisterLoginUserResponse:
        description: RegisterLoginUserResponse represent the response retrived on register or login request.
        schema:
            $ref: '#/definitions/RegisterLoginUserResponseBody'
...
    UpdateUserResponse:
        description: UpdateUserResponse represents response for successfully update user request.
        schema:
            type: string
...
schemes:
    - http
    - https
securityDefinitions:
    bearerAuth:
        description: JWT Authorization header
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"